Flexible generation of JWT tokens
This makes it possible to inject a function for signing
a JWT token with a private key.

The rationale is that libraries like azure keyvault will
take care of storing the private key, but will never actually
deliver the private key but rather does the signing of the
token as part of the SDK. This is more secure as the
app never actually sees the private key then.